Extract edit fade resize rules
